Create Installer.java
This is the installer for periodic work updater. It is designed only for window operating system.
This code takes a installation path, and copies the file from the installer at the location.
Secondly, it also sets up the registry such that the code automatically runs at the start up for the current user.